Extract `AbstractPointer` from `ObjectPointer`
