+++ syft/generic/pointers/object_pointer.py
@@ -1,5 +1,3 @@
-from typing import List
-from typing import Union
 from typing import TYPE_CHECKING
 import weakref
 
@@ -11,20 +9,18 @@ from syft.generic.frameworks.hook.hook_args import register_forward_func
 from syft.generic.frameworks.hook.hook_args import register_backward_func
 from syft.generic.frameworks.hook import hook_args
 from syft.generic.frameworks.types import FrameworkTensor
-from syft.generic.abstract.sendable import AbstractSendable
+from syft.generic.abstract.pointer import AbstractPointer
 from syft.messaging.message import ForceObjectDeleteMessage
 from syft.workers.abstract import AbstractWorker
 
 from syft.exceptions import RemoteObjectFoundError
-from syft.serde.syft_serializable import SyftSerializable
 
 # this if statement avoids circular imports between base.py and pointer.py
 if TYPE_CHECKING:
     from syft.workers.abstract import AbstractWorker
-    from syft.workers.base import BaseWorker
 
 
-class ObjectPointer(AbstractSendable, SyftSerializable):
+class ObjectPointer(AbstractPointer):
     """A pointer to a remote object.
 
     An ObjectPointer forwards all API calls to the remote. ObjectPointer objects
@@ -40,45 +36,6 @@ class ObjectPointer(AbstractSendable, SyftSerializable):
     socket, http, or some other protocol) as that functionality is abstracted
     in the BaseWorker object in self.location.
     """
-
-    def __init__(
-        self,
-        location: "BaseWorker" = None,
-        id_at_location: Union[str, int] = None,
-        owner: "BaseWorker" = None,
-        id: Union[str, int] = None,
-        garbage_collect_data: bool = True,
-        point_to_attr: str = None,
-        tags: List[str] = None,
-        description: str = None,
-    ):
-        """Initializes a ObjectPointer.
-
-        Args:
-            location: An optional BaseWorker object which points to the worker
-                on which this pointer's object can be found.
-            id_at_location: An optional string or integer id of the object
-                being pointed to.
-            owner: An optional BaseWorker object to specify the worker on which
-                the pointer is located. It is also where the pointer is
-                registered if register is set to True. Note that this is
-                different from the location parameter that specifies where the
-                pointer points to.
-            id: An optional string or integer id of the ObjectPointer.
-            garbage_collect_data: If true (default), delete the remote object when the
-                pointer is deleted.
-            point_to_attr: string which can tell a pointer to not point directly to\
-                an object, but to point to an attribute of that object such as .child or
-                .grad. Note the string can be a chain (i.e., .child.child.child or
-                .grad.child.child). Defaults to None, which means don't point to any attr,
-                just point to then object corresponding to the id_at_location.
-        """
-        super().__init__(id=id, owner=owner, tags=tags, description=description)
-
-        self.location = location
-        self.id_at_location = id_at_location
-        self.garbage_collect_data = garbage_collect_data
-        self.point_to_attr = point_to_attr
 
     @staticmethod
     def create_pointer(
